Simplify Vero Names
